% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{class}
\name{BoluspH}
\alias{BoluspH}
\alias{bolusph}
\title{Import and analysis of rumen pH bolus data}
\arguments{
\item{meta_data}{an optional data frame including the columns ID (giving a unique identifier for the cow) and filename (giving the filename used for that cow)}
}
\value{
An object of class BoluspH.
}
\description{
This function is used to import one or more rumen pH bolus data files and store the data for later analysis.  The return value is a data storage class containing all methods required to access, analyse and summarise the data.
}
\details{
The data storage class is written using an OOP-style of programming, so that the data can be accessed in different formats at any point using a single R object as the point of reference.  This minimises the possibilities for programming mistakes by the end user, but does result in code that may look a little strange to users who may be more used to the standard procedural style of R programming.  The best way to understand the code is using an example - see vignette('BoluspH', package='BoluspH') for an overview.
}
\section{Methods}{

\describe{
\item{\code{AddCustomData(folder_path, ID = "all", skip = 1, date_col = 1,
  time_col = 2, pH_col = 3, remove_start = 0.5, remove_end = 0.5,
  max_days = 40, csv = TRUE, sep = ",", dec = ".",
  date_format = "\%Y-\%m-\%d", time_format = "\%H:\%M")}}{Import one or more rumen pH bolus data CSV/Excel files using a custom file format}

\item{\code{AddSmaXtecData(folder_path, ID = "all", remove_start = 0.5,
  remove_end = 0.5, max_days = 40)}}{Import one or more rumen pH bolus data Excel files using the SmaXtec format}

\item{\code{AddWellCowData(folder_path, ID = "all", remove_start = 0.5,
  remove_end = 0.5, max_days = 40, sep = ",", dec = ".")}}{Import one or more rumen pH bolus data CSV files using the WellCow format}

\item{\code{ExtractData(ID = "all", date_limited = TRUE)}}{Extract the saved time series data for one more individual, including model predictions if the analysis has been run}

\item{\code{GetOverview(ID = "all")}}{Extract an overview of the data per animal (assuming that the analysis has been run)}

\item{\code{initialize(meta_data = data.frame(ID = character(0), Filename = character(0)))}}{Set up a BoluspH object (with meta data) before importing files}

\item{\code{RunModels(gam = TRUE, milking = "adjusted", exclude_pre = 1,
  exclude_post = 1, min_days = 10)}}{Run the model using specified options and record the predictions and summary statistics for later use}

\item{\code{StartEndDate(ID, start_date, end_date)}}{Alter the first and last dates to be used for the analysis for a given individual}
}}

\examples{
\dontrun{
vignette('BoluspH', package='BoluspH')
}
}
